+++ CHANGELOG.md
@@ -131,3 +131,4 @@
 - 2023-10-13: Retry deals, session, gp, scheduler, +2
 - 2023-10-14: Pool session, gp, scheduler, qualifier, +2
 - 2023-10-15: Metrics session, deals, qualifier, gp, +3
+- 2023-10-15: Err gp, qualifier, session, deals, +2